+++ bin-proto/src/codec.rs
@@ -21,12 +21,20 @@ pub trait BitDecode<Ctx = (), Tag = ()>: Sized {
 /// Utility functionality for bit-level decoding.
 pub trait BitDecodeExt<Ctx = (), Tag = ()>: BitDecode<Ctx, Tag> {
     /// Parses a new value from its raw byte representation with provided context and tag.
-    fn decode_bytes_ctx<E>(bytes: &[u8], byte_order: E, ctx: &mut Ctx, tag: Tag) -> Result<Self>
+    ///
+    /// Returns a tuple of the parsed value and the number of bits read.
+    fn decode_bytes_ctx<E>(
+        bytes: &[u8],
+        byte_order: E,
+        ctx: &mut Ctx,
+        tag: Tag,
+    ) -> Result<(Self, u64)>
     where
         E: Endianness,
     {
         let mut buffer = BitReader::endian(io::Cursor::new(bytes), byte_order);
-        Self::decode::<_, E>(&mut buffer, ctx, tag)
+        let this = Self::decode::<_, E>(&mut buffer, ctx, tag)?;
+        Ok((this, buffer.position_in_bits()?))
     }
 }
 
@@ -59,21 +67,24 @@ pub trait BitEncodeExt<Ctx = (), Tag = ()>: BitEncode<Ctx, Tag> {
     }
 
     /// Fills the buffer with the raw bytes of this type with provided context and tag.
+    ///
+    /// Returns the number of bytes written
     fn encode_bytes_ctx_buf<E>(
         &self,
         byte_order: E,
         ctx: &mut Ctx,
         tag: Tag,
         buf: &mut [u8],
-    ) -> Result<()>
+    ) -> Result<u64>
     where
         E: Endianness,
     {
-        let mut writer = BitWriter::endian(Cursor::new(buf), byte_order);
+        let mut cursor = Cursor::new(buf);
+        let mut writer = BitWriter::endian(&mut cursor, byte_order);
         self.encode::<_, E>(&mut writer, ctx, tag)?;
         writer.byte_align()?;
 
-        Ok(())
+        Ok(cursor.position())
     }
 }
 
@@ -82,7 +93,9 @@ impl<T, Ctx, Tag> BitEncodeExt<Ctx, Tag> for T where T: BitEncode<Ctx, Tag> {}
 /// A trait with helper functions for simple codecs.
 pub trait BitCodec: BitDecode + BitEncode {
     /// Parses a new value from its raw byte representation without context.
-    fn decode_bytes<E>(bytes: &[u8], byte_order: E) -> Result<Self>
+    ///
+    /// Returns a tuple of the parsed value and the number of bits read
+    fn decode_bytes<E>(bytes: &[u8], byte_order: E) -> Result<(Self, u64)>
     where
         E: Endianness,
     {
@@ -100,7 +113,9 @@ pub trait BitCodec: BitDecode + BitEncode {
     }
 
     /// Fills the buffer with the raw bytes of this type without context.
-    fn encode_bytes_buf<E>(&self, byte_order: E, buf: &mut [u8]) -> Result<()>
+    ///
+    /// Returns the number of bytes written.
+    fn encode_bytes_buf<E>(&self, byte_order: E, buf: &mut [u8]) -> Result<u64>
     where
         E: Endianness,
     {
@@ -201,7 +216,7 @@ macro_rules! test_roundtrip {
             fn roundtrip(x in ::proptest::arbitrary::any::<$ty>()) {
                 use alloc::format; // TODO: https://github.com/proptest-rs/proptest/pull/584
                 let encoded = $crate::BitEncodeExt::encode_bytes_ctx(&x, ::bitstream_io::BigEndian, &mut (), ()).unwrap();
-                let decoded = <$ty as $crate::BitDecodeExt>::decode_bytes_ctx(&encoded, ::bitstream_io::BigEndian, &mut (), ()).unwrap();
+                let decoded = <$ty as $crate::BitDecodeExt>::decode_bytes_ctx(&encoded, ::bitstream_io::BigEndian, &mut (), ()).unwrap().0;
                 ::proptest::prop_assert_eq!(x, decoded);
             }
         );